requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
psycopg2-binary
pandas
scikit-learn
joblib
python-dotenv
pytest
httpx
